+++ Automate/IRecipe.cs
@@ -1,6 +1,6 @@
 using System;
 using StardewValley;
-using Object = StardewValley.Object;
+using SObject = StardewValley.Object;
 
 namespace Pathoschild.Stardew.Automate
 {
@@ -17,7 +17,7 @@ namespace Pathoschild.Stardew.Automate
         int InputCount { get; }
 
         /// <summary>The output to generate (given an input).</summary>
-        Func<Item, Object> Output { get; }
+        Func<Item, SObject> Output { get; }
 
         /// <summary>The time needed to prepare an output.</summary>
         int Minutes { get; }