+++ DataLayers/Layers/Coverage/SprinklerLayer.cs
@@ -7,7 +7,7 @@ using Pathoschild.Stardew.DataLayers.Framework;
 using StardewModdingAPI;
 using StardewValley;
 using StardewValley.TerrainFeatures;
-using Object = StardewValley.Object;
+using SObject = StardewValley.Object;
 
 namespace Pathoschild.Stardew.DataLayers.Layers.Coverage
 {
@@ -17,11 +17,11 @@ namespace Pathoschild.Stardew.DataLayers.Layers.Coverage
         /*********
         ** Fields
         *********/
-        /// <summary>The color for sprinkled tiles.</summary>
-        private readonly Color WetColor = Color.Green;
+        /// <summary>The legend entry for sprinkled tiles.</summary>
+        private readonly LegendEntry Wet;
 
-        /// <summary>The color for unsprinkled tiles.</summary>
-        private readonly Color DryColor = Color.Red;
+        /// <summary>The legend entry for unsprinkled tiles.</summary>
+        private readonly LegendEntry Dry;
 
         /// <summary>The border color for the sprinkler under the cursor.</summary>
         private readonly Color SelectedColor = Color.Blue;
@@ -50,8 +50,8 @@ namespace Pathoschild.Stardew.DataLayers.Layers.Coverage
             this.Mods = mods;
             this.Legend = new[]
             {
-                new LegendEntry(translations.Get("sprinklers.covered"), this.WetColor),
-                new LegendEntry(translations.Get("sprinklers.dry-crops"), this.DryColor)
+                this.Wet = new LegendEntry(translations, "sprinklers.covered", Color.Green),
+                this.Dry = new LegendEntry(translations, "sprinklers.dry-crops", Color.Red)
             };
 
             // get static sprinkler coverage
@@ -78,7 +78,7 @@ namespace Pathoschild.Stardew.DataLayers.Layers.Coverage
 
             // get sprinklers
             Vector2[] searchTiles = visibleArea.Expand(this.MaxRadius).GetTiles().ToArray();
-            Object[] sprinklers =
+            SObject[] sprinklers =
                 (
                     from Vector2 tile in searchTiles
                     where location.objects.ContainsKey(tile)
@@ -90,24 +90,24 @@ namespace Pathoschild.Stardew.DataLayers.Layers.Coverage
 
             // yield sprinkler coverage
             HashSet<Vector2> covered = new HashSet<Vector2>();
-            foreach (Object sprinkler in sprinklers)
+            foreach (SObject sprinkler in sprinklers)
             {
-                TileData[] tiles = this.GetCoverage(sprinkler, sprinkler.TileLocation, coverageBySprinklerID).Select(pos => new TileData(pos, this.WetColor)).ToArray();
+                TileData[] tiles = this.GetCoverage(sprinkler, sprinkler.TileLocation, coverageBySprinklerID).Select(pos => new TileData(pos, this.Wet)).ToArray();
                 foreach (TileData tile in tiles)
                     covered.Add(tile.TilePosition);
-                yield return new TileGroup(tiles, outerBorderColor: sprinkler.TileLocation == cursorTile ? this.SelectedColor : this.WetColor);
+                yield return new TileGroup(tiles, outerBorderColor: sprinkler.TileLocation == cursorTile ? this.SelectedColor : this.Wet.Color);
             }
 
             // yield dry crops
-            TileData[] dryCrops = this.GetDryCrops(location, visibleTiles, covered).Select(pos => new TileData(pos, this.DryColor)).ToArray();
-            yield return new TileGroup(dryCrops, outerBorderColor: this.DryColor);
+            TileData[] dryCrops = this.GetDryCrops(location, visibleTiles, covered).Select(pos => new TileData(pos, this.Dry)).ToArray();
+            yield return new TileGroup(dryCrops, outerBorderColor: this.Dry.Color);
 
             // yield sprinkler being placed
-            Object heldObj = Game1.player.ActiveObject;
+            SObject heldObj = Game1.player.ActiveObject;
             if (heldObj != null && coverageBySprinklerID.ContainsKey(heldObj.ParentSheetIndex))
             {
-                TileData[] tiles = this.GetCoverage(heldObj, cursorTile, coverageBySprinklerID).Select(pos => new TileData(pos, this.WetColor * 0.75f)).ToArray();
-                yield return new TileGroup(tiles, outerBorderColor: this.SelectedColor);
+                TileData[] tiles = this.GetCoverage(heldObj, cursorTile, coverageBySprinklerID).Select(pos => new TileData(pos, this.Wet, this.Wet.Color * 0.75f)).ToArray();
+                yield return new TileGroup(tiles, outerBorderColor: this.SelectedColor, shouldExport: false);
             }
         }
 
@@ -209,8 +209,8 @@ namespace Pathoschild.Stardew.DataLayers.Layers.Coverage
         /// <param name="sprinkler">The sprinkler whose radius to get.</param>
         /// <param name="origin">The sprinkler's tile.</param>
         /// <param name="radii">The sprinkler radii centered on (0, 0) indexed by sprinkler ID.</param>
-        /// <remarks>Derived from <see cref="Object.DayUpdate"/>.</remarks>
-        private IEnumerable<Vector2> GetCoverage(Object sprinkler, Vector2 origin, IDictionary<int, Vector2[]> radii)
+        /// <remarks>Derived from <see cref="SObject.DayUpdate"/>.</remarks>
+        private IEnumerable<Vector2> GetCoverage(SObject sprinkler, Vector2 origin, IDictionary<int, Vector2[]> radii)
         {
             // get relative tiles
             if (!radii.TryGetValue(sprinkler.ParentSheetIndex, out Vector2[] tiles))