+++ DataLayers/release-notes.md
@@ -1,5 +1,6 @@
 # Release notes
 ## Upcoming release
+* Added `data-layers export` console command to export the current data layer to a JSON file.
 * The bee house range now accounts for the Bee House Flower Range Fix mod, if installed.
 * Updated for compatibility with the latest version of Pelican Fiber.
 * Improved translations. Thanks to shiro2579 (updated Portuguese)!